glue_mrpc.py added
